// frontend/src/app/api/models/userRegistrationRequest.ts
/**
 * Montagsmaler API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserRegistrationRequest { 
    username: string;
    email: string;
    password: string;
    firstName?: string;
    lastName?: string;
}